Split out style functions
